WIP: [sveltekit] Merge form and  mutation hooks
