--- packages/sveltekit-integration/src/lib/client/index.ts
@@ -1,2 +1,0 @@
-export * from './client.svelte';
-export * from './mutation.svelte';
